Position Ambiguity Fix
From the spec (aprs101.pdf): The level of ambiguity specified in the latitude will automatically apply to the longitude as well it is not necessary to include any (space) characters in the longitude.

This patch assures that we update lon_min to include a matching number of spaces before processing.  This could be done differently however this approach requires the fewest changes.